requirements.txt
python-dotenv
requests
cloudscraper